+++ src/tycho/core/value_objects/verse.py
@@ -1,0 +1,15 @@
+"""Verse value object."""
+
+from enum import Enum
+
+
+class Verse(Enum):
+    """Enumeration of Verse."""
+
+    FPT = "FPT"
+    FPE = "FPE"
+    FPH = "FPH"
+
+    def __str__(self):
+        """Return string representation."""
+        return self.value